requirements.txt
colormath == 3.0.0
Pillow == 7.2.0
